bastHUD 1.5
Put a lot of time and effort into this one, but it’s finally done. I
changed the main menu and entirely reworked the backpack. I redesigned
the scoreboard, also in minmode. I fixed a lot of bugs including the
spectator-HUD, the Spy’s disguise panel, the targetID and the win
panel. I also changed the stats-summary panel, everything to make the
HUD feel more like one unit.